% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{gauge}
\alias{gauge}
\title{Create elegents interactive gauge illustrations}
\usage{
gauge(value, width = NULL, height = NULL)
}
\arguments{
\item{value}{A numeric between 0 and 100. Represents a percentage.
@param width,height Must be a valid CSS unit (like \code{'100\%'},
 \code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
 string and have \code{'px'} appended.}
}
\description{
An easy way to create gauge charts based on the rstudio
js tutorials with functions for easy implementation in shiny
}
\examples{
gauge(0)

gauge(50)

gauge(1:4)
}
